move serializer to reducer and add action
